Improve: WAV出力部分の切り出し #136
